dont unnecessarily cast

+++ Software/Projects/B-L072Z-LRWAN1/Applications/LoRa/End_Node/LoRaWAN/App/src/bsp.c
@@ -69,7 +69,7 @@ void BSP_sensor_Read( sensor_t *sensor_data)
 	get_location_fix(GPS_LOCATION_FIX_TIMEOUT);
 	#endif
 	
-	uint16_t no_load_solar_voltage = (uint16_t) BSP_GetSolarLevel16();
+	uint16_t no_load_solar_voltage = BSP_GetSolarLevel16();
 	uint16_t load_solar_voltage = get_load_solar_voltage();
